+++ src/rpclib/model/binary.py
@@ -18,11 +18,23 @@
 #
 
 import base64
+import sys
 
-from cStringIO import StringIO
+try:
+    from cStringIO import StringIO
+except ImportError: # Python 3
+    from io import StringIO
+
 from rpclib.model import nillable_string
 from rpclib.model import nillable_iterable
 from rpclib.model import ModelBase
+
+if sys.version > '3':
+    def _join(val):
+        return bytes('').join(val)
+else:
+    def _join(val):
+        return ''.join(val)
 
 class ByteArray(ModelBase):
     """Handles anything other than ascii or unicode-encoded data. Every protocol
@@ -57,12 +69,12 @@ class ByteArray(ModelBase):
     @classmethod
     @nillable_string
     def to_base64(cls, value):
-        return [base64.b64encode(''.join(value))]
+        return [base64.b64encode(_join(value))]
 
     @classmethod
     @nillable_string
     def from_base64(cls, value):
-        return [base64.b64decode(''.join(value))]
+        return [base64.b64decode(_join(value))]
 
 class Attachment(ModelBase):
     """**DEPRECATED!** Use ByteArray instead."""